+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-## 1.7.0
+## 2.1.0
 
-* add more date format - MMMM yyyy, MMMEd, EE, dd MMM yyyy.
+* update dependencies.
+* add more date format - dd EEE, MMM, yy.
+* convert getCurrentDateTime to getCurrentDate function.
+* change getCurrentDateTime function to get full date time.
 
 ## 1.6.0